Move painter image tests to liveMapTest
The Android host JVM has no android.graphics.Bitmap implementation, so tests
that draw a painter to a bitmap cannot run there.
